Master

+++ src/controllers/tutoria_controller.py
@@ -5,15 +5,18 @@ from sqlalchemy.exc import DataError
 from jsonschema.validators import validate
 from jsonschema import ValidationError
 from validate.jsonschema import json_schema_tutoria
+from flask_jwt_extended import jwt_required
+from validate.JWT_validate import JWTValidate
+from Json.jwt_class import JWT
 from flask_weasyprint import HTML, CSS
 import requests
-
 
 tutoria = Blueprint('tutoria', __name__, url_prefix = "/tutoria")
 tutoria_repository = TutoriaRepoImpl()
 tutoria_service = TutoriaService(tutoria_repository)
 
 @tutoria.route("/create_tutoria", methods = ['POST'])
+@jwt_required()
 def create_tutoria():
     
     response = {
@@ -21,6 +24,12 @@ def create_tutoria():
         'message' : 'OK',
         'datos' : []
     }
+
+    current_user = JWT.get_current_user()
+    token = JWTValidate.validar_token_docente(current_user)
+    if token:
+        return jsonify(token)
+
     try:
 
         data = request.get_json()
@@ -53,6 +62,7 @@ def create_tutoria():
         return jsonify(response), 400
 
 @tutoria.route("/tutorias")
+@jwt_required()
 def get_tutorias():
 
     response = {
@@ -61,10 +71,16 @@ def get_tutorias():
         'datos' : []
     }
 
+    current_user = JWT.get_current_user()
+    token = JWTValidate.validar_token_admin(current_user)
+    if token:
+        return jsonify(token)
+
     response['datos'] = tutoria_service.get_tutorias()
     return jsonify(response)
 
 @tutoria.route("/find_tutoria/<string:numero_documento>", methods = ['GET'])
+@jwt_required()
 def get_tutoria_by_docente(numero_documento):
 
     response = {
@@ -72,6 +88,11 @@ def get_tutoria_by_docente(numero_documento):
         'message' : 'OK',
         'datos' : []
     }
+
+    current_user = JWT.get_current_user()
+    token = JWTValidate.validar_token(current_user)
+    if token:
+        return jsonify(token)
 
     tutorias = tutoria_service.find_tutorias_by_docente(numero_documento)
     
@@ -89,6 +110,7 @@ def get_tutoria_by_docente(numero_documento):
         return jsonify(response)
 
 @tutoria.route("/find_tutoria_fecha/<string:fecha>", methods = ['GET'])
+@jwt_required()
 def get_tutoria_by_fecha(fecha):
     
     response = {
@@ -96,6 +118,12 @@ def get_tutoria_by_fecha(fecha):
         'message' : 'OK',
         'datos' : []
     }
+
+    current_user = JWT.get_current_user()
+    token = JWTValidate.validar_token(current_user)
+    if token:
+        return jsonify(token)
+    
     try:        
         tutorias = tutoria_service.find_tutorias_by_fecha(fecha)
 
@@ -112,6 +140,7 @@ def get_tutoria_by_fecha(fecha):
         return jsonify(response), 400
 
 @tutoria.route("/find_tutoria_asignatura/<string:asignatura>", methods = ['GET'])
+@jwt_required()
 def get_tutoria_by_asignatura(asignatura):
 
     response = {
@@ -119,6 +148,12 @@ def get_tutoria_by_asignatura(asignatura):
         'message' : 'OK',
         'datos' : []
     }
+
+    current_user = JWT.get_current_user()
+    token = JWTValidate.validar_token(current_user)
+    if token:
+        return jsonify(token)
+    
     tutorias = tutoria_service.find_tutorias_by_asignatura(asignatura)
 
     if tutorias == None:
@@ -136,12 +171,18 @@ def get_tutoria_by_asignatura(asignatura):
     return jsonify(response)
 
 @tutoria.route("/find_tutoria/<string:documento_docente>/<string:fecha>", methods = ['GET'])
+@jwt_required()
 def get_tutoria_docente_fecha(documento_docente, fecha):
     response = {
         'status_code' : 200,
         'message' : 'OK',
         'datos' : []
     }
+
+    current_user = JWT.get_current_user()
+    token = JWTValidate.validar_token(current_user)
+    if token:
+        return jsonify(token)
 
     tutorias = tutoria_service.find_tutoria_by_docente_fecha(documento_docente, fecha)
     if tutorias:
@@ -153,6 +194,7 @@ def get_tutoria_docente_fecha(documento_docente, fecha):
         return jsonify(response)
 
 @tutoria.route("find_tutoria/<string:fecha>/<string:asignatura>", methods = ['GET'])
+@jwt_required()
 def get_tutoria_fecha_asignatura(fecha, asignatura):
     
     response = {
@@ -160,6 +202,12 @@ def get_tutoria_fecha_asignatura(fecha, asignatura):
         'message' : 'OK',
         'datos' : []
     }
+
+    current_user = JWT.get_current_user()
+    token = JWTValidate.validar_token(current_user)
+    if token:
+        return jsonify(token)
+    
     tutorias = tutoria_service.find_tutoria_by_fecha_asignatura(fecha, asignatura)
 
     if tutorias:
@@ -171,6 +219,7 @@ def get_tutoria_fecha_asignatura(fecha, asignatura):
         return jsonify(response)
     
 @tutoria.route("/find_tutoria/<string:documento_docente>/<string:asignatura>", methods = ['GET'])
+@jwt_required()
 def get_tutoria_docente_asignatura(documento_docente, asignatura):
 
     response = {
@@ -178,6 +227,11 @@ def get_tutoria_docente_asignatura(documento_docente, asignatura):
         'message' : 'OK',
         'datos' : []
     }
+
+    current_user = JWT.get_current_user()
+    token = JWTValidate.validar_token(current_user)
+    if token:
+        return jsonify(token)
 
     tutorias = tutoria_service.find_tutoria_by_docente_asignatura(documento_docente, asignatura)
     if tutorias:
@@ -189,12 +243,18 @@ def get_tutoria_docente_asignatura(documento_docente, asignatura):
         return jsonify(response)
 
 @tutoria.route("/find_tutoria/<string:documento_docente>/<string:fecha>/<string:asignatura>", methods = ['GET'])
+@jwt_required()
 def get_tutoria_docente_fecha_asignatura(documento_docente, fecha, asignatura):
     response = {
         'status_code' : 200,
         'message' : 'OK',
         'datos' : []
     }
+
+    current_user = JWT.get_current_user()
+    token = JWTValidate.validar_token(current_user)
+    if token:
+        return jsonify(token)
 
     tutorias = tutoria_service.find_tutoria_by_docente_fecha_asignatura(documento_docente, fecha, asignatura)
     if tutorias:
@@ -206,12 +266,18 @@ def get_tutoria_docente_fecha_asignatura(documento_docente, fecha, asignatura):
         return jsonify(response)
 
 @tutoria.route("/get_tutorias_soon/<string:documento_docente>", methods = ['GET'])
+@jwt_required()
 def get_tutorias_soon(documento_docente):
     response = {
         'status_code' : 200,
         'message' : 'OK',
         'datos' : []
     }
+
+    current_user = JWT.get_current_user()
+    token = JWTValidate.validar_token_docente(current_user)
+    if token:
+        return jsonify(token)
 
     tutorias = tutoria_service.get_tutorias_soon(documento_docente)
 
@@ -225,6 +291,7 @@ def get_tutorias_soon(documento_docente):
     
 
 @tutoria.route("/get_tutorias_soon_admin", methods = ['GET'])
+@jwt_required()
 def get_tutorias_soon_admin():
     response = {
         'status_code' : 200,
@@ -232,6 +299,11 @@ def get_tutorias_soon_admin():
         'datos' : []
     }
 
+    current_user = JWT.get_current_user()
+    token = JWTValidate.validar_token_admin(current_user)
+    if token:
+        return jsonify(token)
+    
     tutorias = tutoria_service.get_tutorias_soon_admin()
 
     if tutorias:
@@ -243,6 +315,7 @@ def get_tutorias_soon_admin():
         return jsonify(response)
     
 @tutoria.route("/count_tutorias_month_by_docente/<string:documento_docente>", methods = ['GET'])
+@jwt_required()
 def count_tutorias_month_by_docente(documento_docente):
     
     response = {
@@ -250,6 +323,11 @@ def count_tutorias_month_by_docente(documento_docente):
         'message' : 'OK',
         'count' : 0
     }
+
+    current_user = JWT.get_current_user()
+    token = JWTValidate.validar_token_docente(current_user)
+    if token:
+        return jsonify(token)
 
     count_tutorias = tutoria_service.count_tutorias_month_by_docente(documento_docente)
     if count_tutorias:
@@ -260,6 +338,7 @@ def count_tutorias_month_by_docente(documento_docente):
         return jsonify(response)
 
 @tutoria.route("/count_tutorias_month_admin", methods = ['GET'])
+@jwt_required()
 def count_tutorias_month_admin():
     
     response = {
@@ -267,6 +346,11 @@ def count_tutorias_month_admin():
         'message' : 'OK',
         'count' : 0
     }
+
+    current_user = JWT.get_current_user()
+    token = JWTValidate.validar_token_admin(current_user)
+    if token:
+        return jsonify(token)
 
     count_tutorias = tutoria_service.count_tutorias_month_admin()
     if count_tutorias:
@@ -278,6 +362,7 @@ def count_tutorias_month_admin():
 
 
 @tutoria.route("/count_tutorias_week_by_docente/<string:documento_docente>", methods = ['GET'])
+@jwt_required()
 def count_tutorias_week_by_docente(documento_docente):
     
     response = {
@@ -285,6 +370,11 @@ def count_tutorias_week_by_docente(documento_docente):
         'message' : 'OK',
         'count' : 0
     }
+
+    current_user = JWT.get_current_user()
+    token = JWTValidate.validar_token_docente(current_user)
+    if token:
+        return jsonify(token)
 
     count_tutorias = tutoria_service.count_tutorias_week_by_docente(documento_docente)
     if count_tutorias:
@@ -295,6 +385,7 @@ def count_tutorias_week_by_docente(documento_docente):
         return jsonify(response)
     
 @tutoria.route("/count_tutorias_week_admin", methods = ['GET'])
+@jwt_required()
 def count_tutorias_week_admin():
     
     response = {
@@ -302,6 +393,11 @@ def count_tutorias_week_admin():
         'message' : 'OK',
         'count' : 0
     }
+
+    current_user = JWT.get_current_user()
+    token = JWTValidate.validar_token_admin(current_user)
+    if token:
+        return jsonify(token)
 
     count_tutorias = tutoria_service.count_tutorias_week_admin()
     if count_tutorias:
@@ -312,6 +408,7 @@ def count_tutorias_week_admin():
         return jsonify(response)
     
 @tutoria.route("/count_tutorias_day_by_docente/<string:documento_docente>", methods = ['GET'])
+@jwt_required()
 def count_tutorias_day_by_docente(documento_docente):
     
     response = {
@@ -319,6 +416,11 @@ def count_tutorias_day_by_docente(documento_docente):
         'message' : 'OK',
         'count' : 0
     }
+
+    current_user = JWT.get_current_user()
+    token = JWTValidate.validar_token_docente(current_user)
+    if token:
+        return jsonify(token)
 
     count_tutorias = tutoria_service.count_tutorias_day_by_docente(documento_docente)
     if count_tutorias:
@@ -329,6 +431,7 @@ def count_tutorias_day_by_docente(documento_docente):
         return jsonify(response)
     
 @tutoria.route("/count_tutorias_day_admin", methods = ['GET'])
+@jwt_required()
 def count_tutorias_day_admin():
     
     response = {
@@ -336,6 +439,11 @@ def count_tutorias_day_admin():
         'message' : 'OK',
         'count' : 0
     }
+
+    current_user = JWT.get_current_user()
+    token = JWTValidate.validar_token_admin(current_user)
+    if token:
+        return jsonify(token)
 
     count_tutorias = tutoria_service.count_tutorias_day_admin()
     if count_tutorias:
@@ -346,6 +454,7 @@ def count_tutorias_day_admin():
         return jsonify(response)
 
 @tutoria.route("/update_tutoria/<string:id>", methods = ['PUT'])
+@jwt_required()
 def update_tutoria(id):
 
     response = {
@@ -353,6 +462,12 @@ def update_tutoria(id):
         'message' : 'OK',
         'datos' : []
     }
+
+    current_user = JWT.get_current_user()
+    token = JWTValidate.validar_token_docente(current_user)
+    if token:
+        return jsonify(token)
+    
     try:
 
         data = request.get_json()
@@ -382,6 +497,7 @@ def update_tutoria(id):
         return jsonify(response), 400
     
 @tutoria.route("/delete_tutoria/<string:id>", methods = ['DELETE'])
+@jwt_required()
 def delete_tutoria(id):
 
     response = {
@@ -389,6 +505,11 @@ def delete_tutoria(id):
         'message' : 'OK',
         'datos' : []
     }
+
+    current_user = JWT.get_current_user()
+    token = JWTValidate.validar_token_docente(current_user)
+    if token:
+        return jsonify(token)
 
     delete_tutoria = tutoria_service.delete_tutoria(id)
